+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Renamed methods that are supposed to be only used internally:
+  - EveUniverseBaseModel: _esi_mapping()
+  - EveUniverseEntityModel and inherited models: inline_objects(), children(), esi_pk(), has_esi_path_list(), esi_path_list(), esi_path_object(), is_list_only_endpoint()
+
 ### Fixed
 
 ## [0.2.0] - 2020-07-27